# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/analytics/roles": {
      "get": {
        "operationId": "RoleAnalyticsController_getRoles",
        "parameters": [
          {
            "name": "alumniIds",
            "required": false,
            "in": "query",
            "description": "The alumni IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "The start date of the role",
            "schema": {
              "example": "2023-11-16",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "The end date of the role",
            "schema": {
              "example": "2025-04-20",
              "type": "string"
            }
          },
          {
            "name": "facultyIds",
            "required": false,
            "in": "query",
            "description": "The faculty IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "courseIds",
            "required": false,
            "in": "query",
            "description": "The course IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "graduationYears",
            "required": false,
            "in": "query",
            "description": "The graduation years to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "companyIds",
            "required": false,
            "in": "query",
            "description": "The company IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "industryIds",
            "required": false,
            "in": "query",
            "description": "The industry IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "roleCountryCodes",
            "required": false,
            "in": "query",
            "description": "The country codes where alumni exerced their roles",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "roleCityIds",
            "required": false,
            "in": "query",
            "description": "The city IDS where alumni exerced their roles",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "companyHQsCountryCodes",
            "required": false,
            "in": "query",
            "description": "The company HQs country codes to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "companyHQsCityIds",
            "required": false,
            "in": "query",
            "description": "The company HQs city ids to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "currentRolesOnly",
            "required": false,
            "in": "query",
            "description": "Filter for current roles only",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "onlyCompaniesWithSalaryData",
            "required": false,
            "in": "query",
            "description": "Include only companies with salary data",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "onlyInternational",
            "required": false,
            "in": "query",
            "description": "Whether to exclude roles in Portugal",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "excludeResearchAndHighEducation",
            "required": false,
            "in": "query",
            "description": "Exclude research and high education roles",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "alumniSearch",
            "required": false,
            "in": "query",
            "description": "Search query for companies",
            "schema": {
              "example": "search query",
              "type": "string"
            }
          },
          {
            "name": "companySize",
            "required": false,
            "in": "query",
            "description": "The company sizes to filter by",
            "schema": {
              "example": [
                "A",
                "B"
              ],
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "A",
                  "B",
                  "C",
                  "D",
                  "E",
                  "F",
                  "G",
                  "H",
                  "I"
                ]
              }
            }
          },
          {
            "name": "companyType",
            "required": false,
            "in": "query",
            "description": "The company types to filter by",
            "schema": {
              "example": [
                "PUBLIC_COMPANY",
                "PRIVATELY_HELD"
              ],
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "EDUCATIONAL",
                  "GOVERNMENT_AGENCY",
                  "NON_PROFIT",
                  "PARTNERSHIP",
                  "PRIVATELY_HELD",
                  "PUBLIC_COMPANY",
                  "SELF_EMPLOYED",
                  "SELF_OWNED"
                ]
              }
            }
          },
          {
            "name": "escoCodes",
            "required": false,
            "in": "query",
            "description": "The ESCO codes to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "escoClassificationLevel",
            "required": false,
            "in": "query",
            "description": "The ESCO classification level to filter by",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "The number of results to return",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "The offset of the query",
            "schema": {
              "default": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Broad search query",
            "schema": {
              "example": "search query",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "How to sort the results",
            "schema": {
              "default": "count",
              "type": "string",
              "enum": [
                "count",
                "name",
                "year"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "The order of the results",
            "schema": {
              "default": "desc",
              "example": "asc",
              "type": "string"
            }
          },
          {
            "name": "includeTrend",
            "required": true,
            "in": "query",
            "description": "Whether to include the trend data",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The number of roles classified with each ESCO classification, and the level of classification of the role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Returns the number of roles classified with each ESCO classification",
        "tags": [
          "V1"
        ]
      }
    },
...
    "/api/course": {
      "get": {
        "operationId": "CourseController_find",
        "parameters": [
          {
            "name": "courseIds",
            "required": false,
            "in": "query",
            "description": "The courses to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "facultyIds",
            "required": false,
            "in": "query",
            "description": "The faculties to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Returns courses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseAnalyticsEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Get courses",
        "tags": [
          "V1"
        ]
      },
      "post": {
        "operationId": "CourseController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Returns the created course",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseAnalyticsEntity"
                }
              }
            }
          }
        },
        "summary": "Create a course",
        "tags": [
          "V1"
        ]
      }
    },
    "/api/course/{id}": {
      "get": {
        "operationId": "CourseController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Returns a course",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseAnalyticsEntity"
                }
              }
            }
          }
        },
        "summary": "Get a course by id",
        "tags": [
          "V1"
        ]
      }
    },
...
    "/api/user/linkedinAuth": {
      "post": {
        "operationId": "UserController_linkedinAuth",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkedinAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User authenticated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "No matching alumni record found"
          }
        },
        "summary": "Authenticate a user with LinkedIn",
        "tags": [
          "V1"
        ]
      }
    },
    "/api/user/verify-email": {
      "post": {
        "operationId": "UserController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Returns the verified email"
          }
        },
        "summary": "Verify the email of the alumni by generating a token and sending it to the email",
        "tags": [
          "V1"
        ]
      }
    },
    "/api/user/verify-email/token": {
      "post": {
        "operationId": "UserController_verifyEmailToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailTokenDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Returns the status of the validation"
          }
        },
        "summary": "Validate the token against the email",
        "tags": [
          "V1"
        ]
      }
    },
    "/api/user/linkedin-confirm": {
      "post": {
        "operationId": "UserController_linkedinConfirm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkedinAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User authenticated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAuthResponse"
                }
              }
            }
          }
        },
        "summary": "Confirm the LinkedIn profile and generate a JWT token",
        "tags": [
          "V1"
        ]
      }
    },
    "/api/user/check-permission": {
      "post": {
        "operationId": "UserController_checkPermission",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckPermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User has the permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckPermissionResponse"
                }
              }
            }
          }
        },
        "summary": "Check if a user has a permission",
        "tags": [
          "V1"
        ]
      }
    },
    "/api/user/delete-user": {
      "post": {
        "operationId": "UserController_deleteUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User deleted successfully"
          }
        },
        "summary": "Delete a user",
        "tags": [
          "V1"
        ]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DataPointDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "description": "The label of the data point",
            "example": "2021-01-01"
          },
          "value": {
            "type": "number",
            "description": "The value of the data point",
            "example": 100
          }
        },
        "required": [
          "label",
          "value"
        ]
      },
...
      "RoleListItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the role"
          },
          "level": {
            "type": "number",
            "description": "The level of classification of the role, 1-4 being ISCO-08 and 5+ being ESCO"
          },
          "isLeaf": {
            "type": "boolean",
            "description": "Whether the role is a leaf node in the classification tree"
          },
          "code": {
            "type": "string",
            "description": "The ESCO code of the role"
          },
          "count": {
            "type": "number",
            "description": "The number of roles classified with this role"
          },
          "trend": {
            "description": "The trend of the role",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataPointDto"
            }
          }
        },
        "required": [
          "name",
          "level",
          "isLeaf",
          "code",
          "count",
          "trend"
        ]
      },
      "RoleListResponseDto": {
        "type": "object",
        "properties": {
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleListItemDto"
            }
          },
          "count": {
            "type": "number",
            "description": "The total number of roles in the database"
          },
          "filteredCount": {
            "type": "number",
            "description": "The total number of roles in the database after applying the filters"
          },
          "distinctCount": {
            "type": "number",
            "description": "The total number of distinct roles in the database after applying the filters"
          }
        },
        "required": [
          "roles",
          "count",
          "filteredCount",
          "distinctCount"
        ]
      },
...
      "FacultyAnalyticsEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the faculty"
          },
          "name": {
            "type": "string",
            "description": "The name of the faculty"
          },
          "nameInt": {
            "type": "string",
            "description": "The international name of the faculty"
          },
          "acronym": {
            "type": "string",
            "description": "The acronym of the faculty"
          }
        },
        "required": [
          "id",
          "name",
          "nameInt",
          "acronym"
        ]
      },
      "CourseAnalyticsEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the course"
          },
          "name": {
            "type": "string",
            "description": "The name of the course"
          },
          "acronym": {
            "type": "string",
            "description": "The acronym of the course"
          },
          "status": {
            "type": "string",
            "description": "The status of the course",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "courseType": {
            "type": "string",
            "description": "The course type of the course",
            "enum": [
              "INTEGRATED_MASTERS",
              "BACHELORS",
              "MASTERS",
              "DOCTORATE"
            ]
          },
          "facultyAcronym": {
            "type": "string",
            "description": "The faculty acronym of the course"
          },
          "startYear": {
            "type": "number",
            "description": "The start year of the course"
          },
          "endYear": {
            "type": "number",
            "description": "The year in which the course ceased to exist"
          },
          "facultyId": {
            "type": "string",
            "description": "The faculty id of the course"
          },
          "faculty": {
            "description": "The faculty of the course",
            "allOf": [
              {
                "$ref": "#/components/schemas/FacultyAnalyticsEntity"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "acronym",
          "status",
          "courseType",
          "facultyAcronym",
          "startYear",
          "facultyId",
          "faculty"
        ]
      },
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the course",
            "example": "Computer Science"
          },
          "nameInt": {
            "type": "string",
            "description": "The international name of the course",
            "example": "Computer Science"
          },
          "acronym": {
            "type": "string",
            "description": "The acronym of the course",
            "example": "CS"
          },
          "courseType": {
            "type": "string",
            "description": "The type of the course",
            "example": "BACHELORS",
            "enum": [
              "INTEGRATED_MASTERS",
              "BACHELORS",
              "MASTERS",
              "DOCTORATE"
            ]
          },
          "status": {
            "type": "string",
            "description": "The status of the course",
            "example": "ACTIVE",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "facultyId": {
            "type": "string",
            "description": "The faculty id of the course",
            "example": "123"
          },
          "startYear": {
            "type": "number",
            "description": "The start year of the course",
            "example": 2020
          },
          "endYear": {
            "type": "number",
            "description": "If the course is INACTIVE, the end year of the course",
            "example": 2024
          },
          "createdBy": {
            "type": "string",
            "description": "The created by of the course",
            "example": "123"
          }
        },
        "required": [
          "name",
          "acronym",
          "courseType",
          "status",
          "facultyId",
          "startYear"
        ]
      },
...
      "LinkedinAuthDto": {
        "type": "object",
        "properties": {
          "personId": {
            "type": "string",
            "description": "A unique identifying value for the member. Referenced as personId in other API documentation pages."
          },
          "firstName": {
            "type": "string",
            "description": "The first name of the member."
          },
          "lastName": {
            "type": "string",
            "description": "The first name of the member."
          },
          "personalEmail": {
            "type": "string",
            "description": "The personal email of the member."
          },
          "profilePictureUrl": {
            "type": "string",
            "description": "The URL of the member's profile picture."
          },
          "linkedinUrl": {
            "type": "string",
            "description": "The LinkedIn URL of the member."
          }
        },
        "required": [
          "personId",
          "firstName",
          "lastName"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the user"
          },
          "firstName": {
            "type": "string",
            "description": "The first name of the user"
          },
          "lastName": {
            "type": "string",
            "description": "The last name of the user"
          },
          "profilePictureUrl": {
            "type": "string",
            "description": "The profile picture URL of the user"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName"
        ]
      },
      "UserAuthResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "The access token of the user",
            "nullable": true
          },
          "user": {
            "description": "The user object",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "status": {
            "type": "string",
            "description": "The status of the user",
            "enum": [
              "MATCHED",
              "UNMATCHED"
            ]
          },
          "personId": {
            "type": "string",
            "description": "The person ID of the user",
            "nullable": true
          }
        },
        "required": [
          "status"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email of the alumni"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyEmailTokenDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email of the alumni"
          },
          "token": {
            "type": "string",
            "description": "The OTP token of the alumni"
          }
        },
        "required": [
          "email",
          "token"
        ]
      },
      "CheckPermissionDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "The ID of the user to check permission for",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "resource": {
            "type": "string",
            "description": "The resource to check permission for",
            "example": "read"
          },
          "action": {
            "type": "string",
            "description": "The action to check permission for",
            "example": "read"
          }
        },
        "required": [
          "userId",
          "resource",
          "action"
        ]
      },
      "CheckPermissionResponse": {
        "type": "object",
        "properties": {
          "hasPermission": {
            "type": "boolean",
            "description": "Whether the user has permission",
            "example": true
          }
        },
        "required": [
          "hasPermission"
        ]
      },
      "DeleteUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the user to delete"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}